Various Bugfixes and improvements
* Main class definition was added to the maven file
* SQLWrapper will now properly store the config object
* Connection URL is now chosen relative to the dbdriver
* rollback() will no longer do a rollback if autocommit is true
* Escaped newline and tab control characters in DataLoader
implementations
* Fixed wrong SQL command in MySQLDataLoader, "INPATH" was changed to
"INFILE"
* Removed ' from predicate-selection statement because this caused the ?
placeholder to be treated literally
* Fixed a bug where the vp table statistics would contain many NaN
entries. This was fixed by setting the dataset size earlier
* Problem found: mysql maximum identifier length is 64. This is
obviously to short for the tablenames generated by this loader. As a
temporary solution, the "SHORTEN_TABLENAMES" option was introduced. If
this option is set to "true", the program will only create vp tables
with short names