Added property files, jUnit tests and improved query execution
* Application settings are now stored in a .properties file, which is
passed to the application when running from the console.
Additional arguments allow editing the settings which are defined in
this property file
* The query execution time will now be measured in milliseconds. Upon
application start, a timestamp will be printed which indicates when the
application was started.
When execution a query, the query runtime information will still be
printed, but there will also be information as to how much time has
passed between application start and query execution. 
Additionally, the application also prints information about when
(timestamp) the query was started.
* Queries can now be filtered by their name and their position inside
the composite query file. The syntax of the index values works like the
syntax of accessing elements of an array in python.
* Added jUnit tests which test the query filtering algorithm (by index)
and an alternative mod implementation which was required. These tests
will be executed when mvn package gets called